Use BigDecimal, not Float

+++ lib/onix/measure.rb
@@ -3,7 +3,7 @@ module ONIX
     include ROXML
 
     xml_accessor :measure_type_code, :from => "MeasureTypeCode", :as => Fixnum # should be a 2 digit num
-    xml_accessor :measurement,       :from => "Measurement", :as => Float
+    xml_accessor :measurement,       :from => "Measurement", :as => BigDecimal
     xml_accessor :measure_unit_code, :from => "MeasureUnitCode"
   end
 end
